requirements.txt
asgiref==3.7.2
cloudinary==1.34.0
coverage==7.3.1
crispy-bootstrap5==0.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.21
django-active-link==0.1.8
django-allauth==0.56.1
django-crispy-forms==2.0
django-phonenumber-field==7.1.0
django-summernote==0.8.20.0
gunicorn==21.2.0
oauthlib==3.2.2
phonenumbers==8.13.21
Pillow==10.0.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
PyJWT==2.8.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.16
